requirements.txt
colorama==0.4.1
ffmpeg-python==0.1.17
psutil==5.6.2
pymediainfo==4.0

